refactor: extract shared Notion client helpers into _client.py
- new reporting/notion/_client.py owns init_notion_client + format_database_id
  with its own import-time logger
- notion_update / notion_database_structure import and re-export from it,
  dropping the verbatim copies
- editorial imports the two helpers from _client
- addresses finding 1 of #133
