+++ reporting/notion/notion_database_structure.py
@@ -4,13 +4,13 @@ import sys
 import argparse
 from pathlib import Path
 import os
-from notion_client import Client
 import pandas as pd
 from datetime import datetime
 
 # Add the parent directory to sys.path to allow importing from sibling packages
 sys.path.append(str(Path(__file__).parent.parent.parent))
 from config.logger_config import setup_logger
+from reporting.notion._client import init_notion_client, format_database_id
 
 # Set up logger
 logger = None
@@ -21,24 +21,6 @@ def configure_logger(debug_mode=False):
     log_level = logging.DEBUG if debug_mode else logging.INFO
     logger = setup_logger("notion_database_structure", file_logging=False, level=log_level)
     return logger
-
-def init_notion_client(api_token):
-    """Initialize Notion Client using the provided API token."""
-    logger.debug("🔑 Initializing Notion client")
-    try:
-        client = Client(auth=api_token)
-        logger.info("✅ Notion client initialized successfully")
-        return client
-    except Exception as e:
-        logger.error(f"❌ Error initializing Notion client: {e}")
-        return None
-
-def format_database_id(database_id):
-    """Format database ID with hyphens if needed."""
-    if len(database_id) == 32:
-        # Insert hyphens to convert into UUID format
-        return f"{database_id[:8]}-{database_id[8:12]}-{database_id[12:16]}-{database_id[16:20]}-{database_id[20:]}"
-    return database_id
 
 def get_database_structure(notion, database_id):
     """Retrieve the structure of a Notion database."""
